# Improvements to thread safety detection and implement disallowing whole modules