This includes a few things:
- General improvements to thread safety detection like:
	- Call `generic_visit` in all paths and bail out early when already thread unsafe
	- Do not recurse when target is not a function
	- Fix message printed in error report when unsafety is detected because of `__thread_safe__`
- Add ability to blocklist whole modules and do that for `ctypes` and `unittest.mock`

It's easier to review this commit by commit.

Also, here's the test results for `scipy` and `scikit-learn` when it comes to collection performance:

**SciPy**:

|  | Collection time | Collected items to run in parallel |
| --- | --- | --- |
| main   | 33.92s             | 53101 |
| HEAD | 38.19s             | 53089 |

**scikit-learn**:

|  | Collection time | Collected items to run in parallel |
| --- | --- | --- |
| main   | 11.40s             | 23049 |
| HEAD | 10.46s             | 35600 |