@ngoldbaum 2025-06-26
I spent some time profiling this and I didn't see any obvious way to speed things up - now that the visitor is doing more stuff, it takes more time. We also spend a decent amount of time in `ast.parse` and `inspect.getsource` - if there were faster alternatives that would also help, but I don't think there are any right now.