@ngoldbaum 2025-06-26
@rgommers what do you think about the marginally increased collection time for SciPy?